feat: update routes and map path for single equipment
