+++ components/App/Header.vue
@@ -2,25 +2,24 @@
 </script>
 
 <template>
-  <header class="header">
+  <header>
     <NuxtLink to="/">
       <img src="~/assets/aiko.png" alt="logo" />
     </NuxtLink>
     <nav>
       <NuxtLink to="/">Início</NuxtLink>
-      <NuxtLink to="/equipments">Equipamentos</NuxtLink>
       <NuxtLink to="/map">Mapa</NuxtLink>
     </nav>
   </header>
 </template>
 
 <style scoped>
-.header {
+header {
   display: flex;
   align-items: center;
   justify-content: space-between;
   background-color: white;
-  padding: 16px;
+  padding: 16px 32px;
 
   img {
     height: 24px;
